refactor: extract auxiliary sheet builders from _sheet_builder.py (#90)
Move build_assumptions_sheet, build_drivers_sheet, and build_inputs_sheet
into a new _auxiliary_sheets.py module to bring _sheet_builder.py under
the 300-line maximum (383 → 171 lines).
