+++ excel_model/models/comparison.py
@@ -6,11 +6,11 @@ from openpyxl.styles import Alignment
 from excel_model.exceptions import ExcelModelError
 from excel_model.formula_engine import CellContext, render_formula
 from excel_model.loader import InputData
+from excel_model.models._auxiliary_sheets import build_assumptions_sheet
 from excel_model.models._sheet_builder import (
     apply_data_cell_style,
     apply_label_style,
     assign_row_map,
-    build_assumptions_sheet,
     build_model_header,
     group_line_items_by_section,
     write_section_header,
